Make Clang and Clang-Tidy happy.

+++ tests/component/encapsulation.cpp
@@ -277,7 +277,7 @@ TEST(Encapsulation, parseAlternateFormHierarchy) {
     libcellml::Parser parser = libcellml::Parser();
     libcellml::ModelPtr model = parser.parseModel(input);
 
-    EXPECT_EQ(0u, parser.errorCount());
+    EXPECT_EQ(size_t(0), parser.errorCount());
     EXPECT_EQ(1u, model->componentCount());
 }
