+++ tests/parser/file_parser.cpp
@@ -34,7 +34,7 @@ TEST(Parser, parseSineModelFromFile) {
     libcellml::Parser p;
     p.parseModel(buffer.str());
 
-    EXPECT_EQ(0u, p.errorCount());
+    EXPECT_EQ(size_t(0), p.errorCount());
 }
 
 TEST(Parser, parseSineImportsModelFromFile) {
@@ -46,7 +46,7 @@ TEST(Parser, parseSineImportsModelFromFile) {
     libcellml::Parser p;
     p.parseModel(buffer.str());
 
-    EXPECT_EQ(0u, p.errorCount());
+    EXPECT_EQ(size_t(0), p.errorCount());
 }
 
 TEST(Parser, parseInvalidModelFromFile) {
@@ -78,7 +78,7 @@ TEST(Parser, parseOrdModelFromFile) {
     libcellml::Parser p;
     libcellml::ModelPtr model = p.parseModel(buffer.str());
 
-    EXPECT_EQ(0u, p.errorCount());
+    EXPECT_EQ(size_t(0), p.errorCount());
 
     // Test some random values.
     std::string a = model->getComponent("intracellular_ions")->getVariable("BSLmax")->getInitialValue();
@@ -104,7 +104,7 @@ TEST(Parser, parseComplexEncapsulationModelFromFile) {
     libcellml::Parser p;
     p.parseModel(buffer.str());
 
-    EXPECT_EQ(0u, p.errorCount());
+    EXPECT_EQ(size_t(0), p.errorCount());
 }
 
 TEST(Parser, parseModelWithComponentsWithMultipleMathElements) {
@@ -120,7 +120,7 @@ TEST(Parser, parseModelWithComponentsWithMultipleMathElements) {
 
     libcellml::Parser p;
     libcellml::ModelPtr model = p.parseModel(buffer.str());
-    EXPECT_EQ(0u, p.errorCount());
+    EXPECT_EQ(size_t(0), p.errorCount());
 
     std::string a = model->getComponent("c1")->getMath();
     EXPECT_EQ(e1, a);
